+++ owca/allocators.py
@@ -18,6 +18,7 @@ from pprint import pformat
 from typing import List, Dict, Union, Tuple, Optional
 import logging
 
+from owca import platforms
 from owca.logger import trace, TRACE
 from owca.metrics import Metric, MetricType
 from owca.mesos import TaskId
@@ -25,6 +26,8 @@ from owca.platforms import Platform
 from owca.detectors import TasksMeasurements, TasksResources, TasksLabels, Anomaly
 
 from dataclasses import dataclass
+
+from owca.resctrl import _parse_schemata_file_row, check_cbm_bits
 
 log = logging.getLogger(__name__)
 
@@ -36,107 +39,24 @@ class AllocationType(str, Enum):
     RDT = 'rdt'
 
 
-class MergeableAllocationValue(ABC):
+class AllocationValue:
 
     @abstractmethod
-    def merge_with_current(self, current: 'MergeableAllocationValue') -> Tuple[
-            'MergeableAllocationValue', 'MergeableAllocationValue']:
+    def merge_with_current(self, current: 'AllocationValue') -> Tuple[
+        'AllocationValue', Optional['AllocationValue']]:
         ...
-
-
-class SerializableAllocationValue(ABC):
 
     @abstractmethod
     def generate_metrics(self) -> List[Metric]:
         ...
 
-
-@dataclass(unsafe_hash=True)
-class RDTAllocation(SerializableAllocationValue, MergeableAllocationValue):
-    # defaults to TaskId from TasksAllocations
-    name: str = None
-    # CAT: optional - when no provided doesn't change the existing allocation
-    l3: str = None
-    # MBM: optional - when no provided doesn't change the existing allocation
-    mb: str = None
-
-    def generate_metrics(self) -> List[Metric]:
-        """Encode RDT Allocation as metrics.
-        Note:
-        - cache allocation: generated two metrics, with number of cache ways and
-                            mask of bits (encoded as int)
-        - memory bandwidth: is encoded as int, representing MB/s or percentage
-        """
-        # Empty object generate no metric.
-        if not self.l3 and not self.mb:
-            return []
-
-        group_name = self.name or ''
-
-        metrics = []
-        if self.l3:
-            domains = _parse_schemata_file_row(self.l3)
-            for domain_id, raw_value in domains.items():
-                metrics.extend([
-                    Metric(
-                        name='allocation', value=_count_enabled_bits(raw_value),
-                        type=MetricType.GAUGE, labels=dict(
-                            allocation_type='rdt_l3_cache_ways', group_name=group_name,
-                            domain_id=domain_id)
-                    ),
-                    Metric(
-                        name='allocation', value=int(raw_value, 16),
-                        type=MetricType.GAUGE, labels=dict(
-                            allocation_type='rdt_l3_mask', group_name=group_name,
-                            domain_id=domain_id)
-                    )
-                ])
-
-        if self.mb:
-            domains = _parse_schemata_file_row(self.mb)
-            for domain_id, raw_value in domains.items():
-                # NOTE: raw_value is treated as int, ignoring unit used (MB or %)
-                value = int(raw_value)
-                metrics.append(
-                    Metric(
-                       name='allocation', value=value, type=MetricType.GAUGE,
-                       labels=dict(allocation_type='rdt_mb',
-                                   group_name=group_name, domain_id=domain_id)
-                    )
-                )
-
-        return metrics
-
-    @trace(log)
-    def merge_with_current(self: Optional['RDTAllocation'],
-                           current_rdt_allocation: 'RDTAllocation') -> \
-            Tuple['RDTAllocation', 'RDTAllocation']:
-        """Merge with existing RDTAllocation objects and return
-        sum of the allocations (target_rdt_allocation) and allocations that need to be updated
-        (rdt_allocation_changeset)."""
-        new_rdt_allocation = self
-        # new name, then new allocation will be used (overwrite) but no merge
-        if current_rdt_allocation is None or current_rdt_allocation.name != new_rdt_allocation.name:
-            log.debug('new name or no previous allocation exists')
-            return new_rdt_allocation, new_rdt_allocation
-        else:
-            log.debug('merging existing rdt allocation')
-            target_rdt_allocation = RDTAllocation(
-                name=current_rdt_allocation.name,
-                l3=new_rdt_allocation.l3 or current_rdt_allocation.l3,
-                mb=new_rdt_allocation.mb or current_rdt_allocation.mb,
-            )
-            rdt_allocation_changeset = RDTAllocation(
-                name=new_rdt_allocation.name,
-            )
-            if current_rdt_allocation.l3 != new_rdt_allocation.l3:
-                rdt_allocation_changeset.l3 = new_rdt_allocation.l3
-            if current_rdt_allocation.mb != new_rdt_allocation.mb:
-                rdt_allocation_changeset.mb = new_rdt_allocation.mb
-            return target_rdt_allocation, rdt_allocation_changeset
+    @abstractmethod
+    def validate(self) -> List[str]:
+        """Returns list of errors, empty list indicates that value is ok."""
+        ...
 
 
-TaskAllocations = Dict[AllocationType, Union[float, RDTAllocation]]
+TaskAllocations = Dict[AllocationType, Union[float, AllocationValue]]
 TasksAllocations = Dict[TaskId, TaskAllocations]
 
 
@@ -195,7 +115,7 @@ def _convert_tasks_allocations_to_metrics(tasks_allocations: TasksAllocations) -
         task_allocations: TaskAllocations
         for allocation_type, allocation_value in task_allocations.items():
 
-            if isinstance(allocation_value, SerializableAllocationValue):
+            if isinstance(allocation_value, AllocationValue):
                 this_allocation_metrics = allocation_value.generate_metrics()
                 for metric in this_allocation_metrics:
                     metric.labels.update(task_id=task_id)
@@ -234,15 +154,14 @@ def _calculate_task_allocations_changeset(
     task_allocations_changeset: TaskAllocations = {}
 
     for allocation_type, new_allocation_value in new_task_allocations.items():
-        if isinstance(new_allocation_value, MergeableAllocationValue):
-            current_rdt_allocations = current_task_allocations.get(AllocationType.RDT)
-            target_rdt_allocation, rdt_allocation_changeset = \
-                new_allocation_value.merge_with_current(current_rdt_allocations)
-            target_rdt_allocation: RDTAllocation
-            rdt_allocation_changeset: RDTAllocation
-            target_task_allocations[AllocationType.RDT] = target_rdt_allocation
-            if rdt_allocation_changeset.l3 is not None or rdt_allocation_changeset.mb is not None:
-                task_allocations_changeset[AllocationType.RDT] = rdt_allocation_changeset
+        if isinstance(new_allocation_value, AllocationValue):
+            current_allocation = current_task_allocations.get(allocation_type)
+            target_allocation, allocation_changeset = \
+                new_allocation_value.merge_with_current(current_allocation)
+            target_task_allocations[allocation_type] = target_allocation
+
+            if allocation_changeset is not None:
+                task_allocations_changeset[allocation_type] = allocation_changeset
 
         else:
             # Float and integered based change detection.
@@ -320,56 +239,24 @@ def _calculate_tasks_allocations_changeset(
     return target_tasks_allocations, tasks_allocations_changeset
 
 
-def _parse_schemata_file_row(line: str) -> Dict[str, str]:
-    """Parse RDTAllocation.l3 and RDTAllocation.mb strings based on
-    https://elixir.bootlin.com/linux/latest/source/arch/x86/kernel/cpu/intel_rdt_ctrlmondata.c#lL206
-    and return dict mapping and domain id to its configuration (value).
-    Resource type (e.g. mb, l3) is dropped.
-
-    Eg.
-    mb:1=20;2=50 returns {'1':'20', '2':'50'}
-    mb:xxx=20mbs;2=50b returns {'1':'20mbs', '2':'50b'}
-    raises ValueError exception for inproper format or conflicting domains ids.
+def _ignore_invalid_allocations(platform: platforms.Platform,
+                                new_tasks_allocations: TasksAllocations) -> (
+        int, TasksAllocations):
+    """Validate and ignore allocations, that are invaild.
+    Returns new valid TasksAllocation object and number of ignored allocations.
     """
-    RESOURCE_ID_SEPARATOR = ':'
-    DOMAIN_ID_SEPARATOR = ';'
-    VALUE_SEPARATOR = '='
+    # Ignore and warn about invalid allocations.
+    ignored_allocations = 0
+    task_ids_to_remove = set()
+    for task_id, task_allocations in new_tasks_allocations.items():
 
-    domains = {}
-
-    # Ignore emtpy line.
-    if not line:
-        return {}
-
-    # Drop resource identifier prefix like ("mb:")
-    line = line[line.find(RESOURCE_ID_SEPARATOR)+1:]
-    # Domains
-    domains_with_values = line.split(DOMAIN_ID_SEPARATOR)
-    for domain_with_value in domains_with_values:
-        if not domain_with_value:
-            raise ValueError('domain cannot be empty')
-        if VALUE_SEPARATOR not in domain_with_value:
-            raise ValueError('Value separator is missing "="!')
-        separator_position = domain_with_value.find(VALUE_SEPARATOR)
-        domain_id = domain_with_value[:separator_position]
-        if not domain_id:
-            raise ValueError('domain_id cannot be empty!')
-        value = domain_with_value[separator_position+1:]
-        if not value:
-            raise ValueError('value cannot be empty!')
-
-        if domain_id in domains:
-            raise ValueError('Conflicting domain id found!')
-
-        domains[domain_id] = value
-
-    return domains
+        if isinstance(task_allocations, AllocationValue):
+            errors = task_allocations.validate()
 
 
-def _count_enabled_bits(hexstr: str) -> int:
-    """Parse a raw value like f202 to number of bits enabled."""
-    if hexstr == '':
-        return 0
-    value_int = int(hexstr, 16)
-    enabled_bits_count = bin(value_int).count('1')
-    return enabled_bits_count
+
+    new_tasks_allocations = {t: a for t, a in new_tasks_allocations.items()
+                             if t not in task_ids_to_remove}
+    return ignored_allocations, new_tasks_allocations
+
+
